refactor: extract BibliographySettings panel from SettingsDialog (#672)
Fourth per-tab panel split. The bibliography tab (#302) — the CSL citation
style picker + the imported user styles/locales lists with import/remove — is
self-contained on api.bibliography.* / api.csl.*, so it lifts whole into
BibliographySettings.svelte. SettingsDialog 1640 → 1363. With both skills and
bibliography now extracted, the dialog also sheds the orphaned .csl-* /
.hint.empty / .csl-error CSS.

Tests (7): components/BibliographySettings.test.ts mocks api.bibliography +
api.csl and pins the on-mount load + current-style selection, the empty states,
the style change → setStyle, the imported lists, Import → import+reload, Remove
→ removeStyle(id), and the import-error banner.

Verified: pnpm lint (0 errors), full suite (3005 passed, +7), electron-forge
build + boot smoke.
